+++ controllers/app/account.js
@@ -1,0 +1,3 @@
+exports.getAccount = (req, res) => {
+  res.render("account", { selected: "Mot de Passe Oublié" });
+};
