+++ controllers/app/error.js
@@ -1,0 +1,5 @@
+exports.getError = (req, res) => {
+  res.status(404);
+
+  res.render("error", { selected: "Erreur" });
+};
