+++ controllers/app/forgot-password.js
@@ -1,0 +1,3 @@
+exports.getForgotPassword = (req, res) => {
+  res.render("forgot-password", { selected: "Mot de Passe Oublié" });
+};
