+++ controllers/app/home.js
@@ -1,0 +1,3 @@
+exports.getHome = (req, res) => {
+    res.render("home", { selected: "Accueil" });
+  }
