+++ controllers/app/sign-in-up.js
@@ -1,0 +1,7 @@
+exports.getSignin = (req, res) => {
+  res.render("signin", { selected: "Connexion" });
+};
+
+exports.getSignup = (req, res) => {
+  res.render("signup", { selected: "Inscription" });
+};
