+++ controllers/app/watchlists.js
@@ -1,0 +1,6 @@
+exports.getWatchlists = (req, res) => {
+    res.render("my-watchlist", {
+      selected: "Ma Watchlist",
+      user: req.session.user,
+    });
+  }
