+++ views/settings.ejs
@@ -54,7 +54,7 @@
             </div>
             <div class="div-container">
                 <h1 class="title">Bio</h1>
-                <textarea name="Bio" class="settings-bio" maxlength="255" id="bio"><%=user.bio%></textarea>
+                <textarea name="newbio" class="settings-bio" maxlength="255" id="bio"><%=user.bio%></textarea>
                 <button type="submit" onclick="ModifyBio()">Modifier la Bio</button>
             </div>
